settings removal and feature merged to display
