+++ admin/src/js/main.js
@@ -37,7 +37,6 @@ require('./controllers/images-branding');
 require('./controllers/images-partners');
 require('./controllers/import-translation');
 require('./controllers/message-queue');
-require('./controllers/settings');
 require('./controllers/sms-forms');
 require('./controllers/sms-settings');
 require('./controllers/sms-test');
@@ -198,11 +197,6 @@ angular.module('adminApp').config(function(
           templateUrl: 'templates/display_translations.html'
         }
       }
-    })
-    .state('settings', {
-      url: '/settings',
-      controller: 'SettingsCtrl',
-      templateUrl: 'templates/settings.html'
     })
     .state('backup', {
       url: '/backup',
